Spin Map into its own component
